+++ CHANGELOG.md
@@ -22,6 +22,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Replaced `unwrap()` in `sensors`.
 - Replaced remaining `unwrap()` (mostly in `main`)
 - Replaced `.expect()` in `multi_driver` with an `unreachable!()`
+- Removed `.expect()` in `multi_driver` and returned the error instead.
+- Removed `.expect()` in `rtc` and returned the rror instead.
 
 ## [1.1.5] - 2026-07-04